backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
